requirements.txt
torch>=2.0.0
torchtext>=0.15.0
sacrebleu>=2.3.1
sentencepiece>=0.1.99
tokenizers>=0.13.3
datasets>=2.14.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
